chore: move apply-auth-mode, construct-exeInfo, get-amplify-appId
