+++ packages/amplify-cli-core/src/extensions/get-project-meta.ts
@@ -1,4 +1,4 @@
-import { $TSMeta, projectNotInitializedError, stateManager } from 'amplify-cli-core';
+import { $TSMeta, projectNotInitializedError, stateManager } from '..';
 
 /**
  * returns project meta
